+++ SharePointFramework/PortfolioWebParts/src/components/ProjectTimeline/index.tsx
@@ -1,118 +1,25 @@
-import { get } from '@microsoft/sp-lodash-subset'
 import { getId } from '@uifabric/utilities'
-import sortArray from 'array-sort'
-import { ITimelineData, ITimelineGroup, ITimelineItem } from 'interfaces'
-import moment from 'moment'
 import { CommandBar, ICommandBarProps } from 'office-ui-fabric-react/lib/CommandBar'
 import { ContextualMenuItemType } from 'office-ui-fabric-react/lib/ContextualMenu'
-import { IColumn } from 'office-ui-fabric-react/lib/DetailsList'
 import { MessageBar, MessageBarType } from 'office-ui-fabric-react/lib/MessageBar'
 import { format } from 'office-ui-fabric-react/lib/Utilities'
 import { Spinner } from 'office-ui-fabric-react/lib/Spinner'
 import * as strings from 'PortfolioWebPartsStrings'
-import React, { Component } from 'react'
+import React, { FunctionComponent } from 'react'
 import _ from 'underscore'
-import { FilterPanel, IFilterItemProps, IFilterProps } from '../FilterPanel'
+import { FilterPanel } from '../FilterPanel'
 import { DetailsCallout } from './DetailsCallout'
 import { Timeline } from './Timeline'
 import styles from './ProjectTimeline.module.scss'
-import { IProjectTimelineProps, IProjectTimelineState } from './types'
-import { ProjectListModel, TimelineContentListModel } from 'models'
+import { IProjectTimelineProps } from './types'
 import './ProjectTimeline.overrides.css'
 import { UserMessage } from 'pp365-shared/lib/components/UserMessage'
+import { useProjectTimeline } from './useProjectTimeline'
+import { ProjectTimelineContext } from './context'
+import { ITimelineItem } from 'interfaces'
 
-/**
- * @component ProjectTimeline
- * @extends Component
- */
-export class ProjectTimeline extends Component<IProjectTimelineProps, IProjectTimelineState> {
-  /**
-   * Constructor
-   *
-   * @param props Props
-   */
-  constructor(props: IProjectTimelineProps) {
-    super(props)
-    this.state = { loading: true, showFilterPanel: false, activeFilters: {} }
-    moment.locale('nb')
-  }
-
-  public async componentDidMount(): Promise<void> {
-    try {
-      const [data, timelineConfiguration] = await this._fetchData()
-      this.setState({ data, timelineConfiguration, loading: false })
-    } catch (error) {
-      this.setState({ error: error.message || error.status, loading: false })
-    }
-  }
-
-  public render(): React.ReactElement<IProjectTimelineProps> {
-    if (this.state.loading) {
-      return (
-        <div className={styles.root}>
-          <div className={styles.container}>
-            <Spinner label={format(strings.LoadingText, this.props.title)} />
-          </div>
-        </div>
-      )
-    }
-    if (this.state.error) {
-      return (
-        <div className={styles.root}>
-          <div className={styles.container}>
-            <UserMessage text={this.state.error} type={MessageBarType.error} />
-          </div>
-        </div>
-      )
-    }
-
-    const { groups, items } = this._getFilteredData()
-
-    return (
-      <div className={styles.root}>
-        <div className={styles.container}>
-          <div className={styles.commandBar}>
-            <CommandBar {...this._getCommandBarProps()} />
-          </div>
-          <div className={styles.header}>
-            <div className={styles.title}>{this.props.title}</div>
-          </div>
-          <div className={styles.infoText}>
-            <MessageBar>
-              <div
-                dangerouslySetInnerHTML={{
-                  __html: format(
-                    this.props.infoText ? this.props.infoText : strings.ProjectTimelineInfoText,
-                    encodeURIComponent(window.location.href)
-                  )
-                }}></div>
-            </MessageBar>
-          </div>
-          <Timeline
-            defaultTimeStart={[-1, 'months']}
-            defaultTimeEnd={[1, 'years']}
-            _onItemClick={this._onItemClick.bind(this)}
-            groups={groups}
-            items={items}
-          />
-        </div>
-        <FilterPanel
-          isOpen={this.state.showFilterPanel}
-          headerText={strings.FilterText}
-          filters={this._getFilters()}
-          onFilterChange={this._onFilterChange.bind(this)}
-          isLightDismiss
-          onDismiss={() => this.setState({ showFilterPanel: false })}
-        />
-        {this.state.showDetails && (
-          <DetailsCallout
-            timelineItem={this.state.showDetails}
-            onDismiss={() => this.setState({ showDetails: null })}
-          />
-        )}
-      </div>
-    )
-  }
+export const ProjectTimeline: FunctionComponent<IProjectTimelineProps> = (props) => {
+  const { state, setState, onFilterChange } = useProjectTimeline(props)
 
   /**
    * On item click
@@ -120,90 +27,13 @@ export class ProjectTimeline extends Component<IProjectTimelineProps, IProjectTi
    * @param event Event
    * @param item Item
    */
-  private _onItemClick(event: React.MouseEvent<HTMLDivElement, MouseEvent>, item: ITimelineItem) {
-    this.setState({ showDetails: { element: event.currentTarget, item } })
+  const onItemClick = (event: React.MouseEvent<HTMLDivElement, MouseEvent>, item: ITimelineItem) => {
+    setState({ showDetails: { element: event.currentTarget, item } })
   }
-
-  /**
-   * Get filtered data
-   */
-  private _getFilteredData(): ITimelineData {
-    const { activeFilters, data } = { ...this.state } as IProjectTimelineState
-    const activeFiltersKeys = Object.keys(activeFilters)
-    data.items = sortArray(data.items, 'data.sortOrder')
-
-    const projectId = data.items.find(
-      (i) => i?.projectUrl === this.props.pageContext.site.absoluteUrl
-    )?.id
-    const topGroup = data.groups.find((i) => i?.id === projectId)
-    projectId &&
-      (data.groups = [topGroup, ...data.groups.filter((grp) => grp?.id !== projectId)].filter(
-        (grp) => grp
-      ))
-
-    if (activeFiltersKeys.length > 0) {
-      const items = activeFiltersKeys.reduce(
-        (newItems, key) => newItems.filter((i) => activeFilters[key].indexOf(get(i, key)) !== -1),
-        data.items
-      )
-      const groups = data.groups.filter((grp) => items.filter((i) => i.group === grp.id).length > 0)
-      return { items, groups }
-    } else {
-      return data
-    }
-  }
-
-  /**
-   * Get filters
-   */
-  private _getFilters(): IFilterProps[] {
-    const config = this.state.timelineConfiguration
-    const columns = [
-      config.find((item) => item?.Title === strings.ProjectLabel).GtTimelineFilter && {
-        fieldName: 'project',
-        name: strings.SiteTitleLabel,
-        isCollapsed: true
-      },
-      { fieldName: 'data.type', name: strings.TypeLabel },
-      { fieldName: 'data.tag', name: strings.TagFieldLabel }
-    ]
-    const hiddenItems = config.filter((item) => !item?.GtTimelineFilter).map((item) => item.Title)
-
-    return columns.map((col) => ({
-      column: { key: col.fieldName, minWidth: 0, ...col },
-      items: this.state.data.items
-        .filter((item) => !hiddenItems.includes(item.data?.type))
-        .map((i) => get(i, col.fieldName))
-        .filter((value, index, self) => value && self.indexOf(value) === index)
-        .map((name) => {
-          const filter = this.state.activeFilters[col.fieldName]
-          const selected = filter ? filter.indexOf(name) !== -1 : false
-          return { name, value: name, selected }
-        }),
-      defaultCollapsed: col.isCollapsed
-    }))
-  }
-
-  /**
-   * On filter change
-   *
-   * @param column Column
-   * @param selectedItems Selected items
-   */
-  private _onFilterChange(column: IColumn, selectedItems: IFilterItemProps[]) {
-    const { activeFilters } = { ...this.state } as IProjectTimelineState
-    if (selectedItems.length > 0) {
-      activeFilters[column.fieldName] = selectedItems.map((i) => i.value)
-    } else {
-      delete activeFilters[column.fieldName]
-    }
-    this.setState({ activeFilters })
-  }
-
   /**
    * Get command bar items
    */
-  private _getCommandBarProps(): ICommandBarProps {
+  const getCommandBarProps = (): ICommandBarProps => {
     const cmd: ICommandBarProps = { items: [], farItems: [] }
     cmd.farItems.push({
       key: getId('Filter'),
@@ -214,172 +44,70 @@ export class ProjectTimeline extends Component<IProjectTimelineProps, IProjectTi
       iconOnly: true,
       onClick: (ev) => {
         ev.preventDefault()
-        this.setState({ showFilterPanel: true })
+        setState({ showFilterPanel: true })
       }
     })
     return cmd
   }
 
-  /**
-   * Creating groups based on projects title
-   *
-   * @param projects Projects
-   *
-   * @returns Timeline groups
-   */
-  private _transformGroups(projects: ProjectListModel[]): ITimelineGroup[] {
-    const mappedProjects = _.uniq(projects.map((project) => project.title)).map((title) => {
-      const project = projects.find((project) => project.title === title)
-      return {
-        title: project.title,
-        siteId: project.siteId
-      }
-    })
-
-    const groups: ITimelineGroup[] = mappedProjects.map((project, id) => {
-      return {
-        id,
-        title: project.title,
-        siteId: project.siteId
-      }
-    })
-    return sortArray(groups, ['type', 'title'])
-  }
-
-  /**
-   * Transform items for timeline
-   *
-   * @param timelineItems Timeline items
-   * @param groups Groups
-   *
-   * @returns Timeline items
-   */
-  private _transformItems(
-    timelineItems: TimelineContentListModel[],
-    groups: ITimelineGroup[]
-  ): ITimelineItem[] {
-    let _item, _siteId
-    try {
-      const items: ITimelineItem[] = timelineItems.map((item, id) => {
-        _item = item
-
-        const group = _.find(groups, (grp) => item.siteId.indexOf(grp.siteId) !== -1)
-        _siteId = group.siteId || 'N/A'
-
-        if (group === null) return
-
-        const style: React.CSSProperties = {
-          color: 'white',
-          border: 'none',
-          cursor: 'auto',
-          outline: 'none',
-          background:
-            item.elementType !== strings.BarLabel ? 'transparent' : item.hexColor || '#f35d69',
-          backgroundColor:
-            item.elementType !== strings.BarLabel ? 'transparent' : item.hexColor || '#f35d69'
-        }
-        return {
-          id,
-          group: group.id,
-          title:
-            item.type === strings.ProjectLabel
-              ? format(strings.ProjectTimelineItemInfo, item.title)
-              : item.itemTitle,
-          start_time:
-            item.elementType !== strings.BarLabel
-              ? moment(new Date(item.endDate))
-              : moment(new Date(item.startDate)),
-          end_time: moment(new Date(item.endDate)),
-          itemProps: { style },
-          project: item.title,
-          projectUrl: item.url,
-          data: {
-            phase: item.phase,
-            description: item.description,
-            type: item.type,
-            budgetTotal: item.budgetTotal,
-            costsTotal: item.costsTotal,
-            sortOrder: item.sortOrder,
-            hexColor: item.hexColor,
-            elementType: item.elementType,
-            filter: item.timelineFilter,
-            tag: item.tag
-          }
-        } as ITimelineItem
-      })
-
-      return items.filter((i) => i)
-    } catch (error) {
-      throw new Error(
-        format(
-          strings.ProjectTimelineErrorTransformItemText,
-          _siteId,
-          _item.itemTitle ? `${_item.itemTitle} (${_item.title})` : _item.title,
-          _item.type,
-          error
-        )
-      )
-    }
-  }
-
-  /**
-   * Fetch data
-   *
-   * @returns Timeline data and timeline configuration
-   */
-  private async _fetchData(): Promise<[ITimelineData, any]> {
-    try {
-      const timelineConfiguration = await this.props.dataAdapter.fetchTimelineConfiguration()
-
-      const [
-        projects,
-        projectData,
-        timelineContentItems,
-        timelineAggregatedContent = []
-      ] = await Promise.all([
-        this.props.dataAdapter.fetchEnrichedProjects(),
-        this.props.dataAdapter.fetchTimelineProjectData(timelineConfiguration),
-        this.props.dataAdapter.fetchTimelineContentItems(timelineConfiguration),
-        this.props.dataAdapter.fetchTimelineAggregatedContent(
-          this.props.configItemTitle,
-          this.props.dataSourceName,
-          timelineConfiguration
-        )
-      ])
-
-      const filteredProjects = projects.filter((project) => {
-        return project.startDate !== null && project.endDate !== null
-      })
-
-      const filteredTimelineItems = [...timelineContentItems, ...timelineAggregatedContent].filter(
-        (item) => {
-          return filteredProjects.some((project) => {
-            return project.title.indexOf(item.title) !== -1
-          })
-        }
-      )
-
-      let timelineItems: TimelineContentListModel[] = filteredProjects.map((project) => {
-        const config = projectData.configElement
-        const statusReport = projectData.reports.find((statusReport) => {
-          return statusReport.siteId === project.siteId
-        })
-        return {
-          ...project,
-          ...statusReport,
-          ...config
-        }
-      })
-
-      timelineItems = [...timelineItems, ...filteredTimelineItems]
-      const groups = this._transformGroups(filteredProjects)
-      const items = this._transformItems(timelineItems, groups)
-
-      return [{ items, groups }, timelineConfiguration]
-    } catch (error) {
-      throw error
-    }
-  }
+  return (
+    <ProjectTimelineContext.Provider value={{ props, state, setState, onFilterChange }}>
+      <div className={styles.root}>
+        <div className={styles.container}>
+          <div className={styles.commandBar}>
+            <CommandBar {...getCommandBarProps()} />
+          </div>
+          <div className={styles.header}>
+            <div className={styles.title}>{props.title}</div>
+          </div>
+          <div className={styles.infoText}>
+            <MessageBar>
+              <div
+                dangerouslySetInnerHTML={{
+                  __html: format(
+                    props.infoText ? props.infoText : strings.ProjectTimelineInfoText,
+                    encodeURIComponent(window.location.href)
+                  )
+                }}></div>
+            </MessageBar>
+          </div>
+          {state.loading ? (
+            <div className={styles.root}>
+              <div className={styles.container}>
+                <Spinner label={format(strings.LoadingText, props.title)} />
+              </div>
+            </div>
+          ) :
+            state.error ? (
+              <UserMessage type={MessageBarType.severeWarning} text={state.error.message} />
+            ) : (
+              <>
+                <Timeline
+                  defaultTimeStart={[-1, 'months']}
+                  defaultTimeEnd={[1, 'years']}
+                  _onItemClick={onItemClick.bind(this)}
+                  groups={state.filteredData.groups}
+                  items={state.filteredData.items}
+                />
+                <FilterPanel
+                  isOpen={state.showFilterPanel}
+                  headerText={strings.FilterText}
+                  filters={state.filters}
+                  onFilterChange={onFilterChange.bind(this)}
+                  isLightDismiss
+                  onDismiss={() => setState({ showFilterPanel: false })}
+                />
+              </>)}
+        </div>
+        {state.showDetails && (
+          <DetailsCallout
+            timelineItem={state.showDetails}
+            onDismiss={() => setState({ showDetails: null })}
+          />
+        )}
+      </div>
+    </ProjectTimelineContext.Provider>
+  )
 }
 
-export { IProjectTimelineProps }
+export * from './types'
